requirements.txt
# requirements.txt

# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
langchain>=0.0.300
langchain-together>=0.0.2.post1  # Together AI integration
together>=0.2.8           # Together AI client
numpy>=1.24.3            # Compatible with Python 3.11
pandas>=2.0.3            # Compatible with Python 3.11

# Database and async support
databases>=0.5.2
asyncio>=3.4.3
aiohttp>=3.8.1

# Translation
azure-ai-translation-text>=1.0.0

# Utilities
python-multipart>=0.0.6
pydantic>=1.8.2
typing-extensions>=3.10.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
pytest-cov>=4.1.0

# Logging and monitoring
flake8>=3.9.2
mypy>=0.910
requests>=2.31.0